implement chatgpt recommend service

+++ src/main/java/com/example/rb/recommend/InputRepository.java
@@ -2,5 +2,5 @@ package com.example.rb.recommend;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InputRepository extends JpaRepository<Input, Long> {
+public interface InputRepository extends JpaRepository<Input, Integer> {
 }
